added inline style
I found a random stack overflow post here: https://stackoverflow.com/questions/17870793/parseint-is-returning-nan-for-css-property that ascertained usage of things such as 'style.left'and 'style.top'are specifically for inline styling. so at line 51 i added a top style for my animate codeblock. It would be better if this was in the css later on but for now i will leave it like this